ExperimentList - A container for multi-experiment data

The ExperimentList class is a container that builds on the SimpleList with additional checks for consistency in experiment names and length. It contains a SimpleList of experiments with sample identifiers. One element present per experiment performed.

Convert from SimpleList or list to the multi-experiment data container. When using the mergeReplicates method, additional arguments are passed to the given simplify function argument (e.g., na.rm = TRUE)

Usage

# S4 method for class 'ExperimentList'
show(object)

# S4 method for class 'ExperimentList'
isEmpty(x)

# S4 method for class 'ExperimentList'
dimnames(x)

# S4 method for class 'ExperimentList'
colnames(x, do.NULL = TRUE, prefix = "col")

# S4 method for class 'ExperimentList'
rownames(x, do.NULL = TRUE, prefix = "row")

# S4 method for class 'ExperimentList'
mergeReplicates(x, replicates = list(), simplify = BiocGenerics::mean, ...)

# S4 method for class 'ANY,missing'
assay(x, i, withDimnames = TRUE, ...)

# S4 method for class 'ExperimentList'
assays(x, withDimnames = TRUE, ...)

# S4 method for class 'ExperimentList,missing'
assay(x, i, withDimnames = TRUE, ...)

# S4 method for class 'ExperimentList,numeric'
assay(x, i, withDimnames = TRUE, ...)

# S4 method for class 'ExperimentList,character'
assay(x, i, withDimnames = TRUE, ...)

Arguments

object, x

An ExperimentList object

do.NULL, prefix

See ?base::rownames for a description of these arguments.

replicates

mergeReplicates: A list or LogicalList where each element represents a sample and a vector of repeated measurements for the sample

simplify

A function for merging columns where duplicates are indicated by replicates

...

Additional arguments. See details for more information.

i

A scalar character or integer index

withDimnames

logical (default TRUE) whether to return dimension names

Value

An ExperimentList class object

Methods (by generic)

show(ExperimentList): Show method for ExperimentList class

isEmpty(ExperimentList): check for zero length across all experiments

dimnames(ExperimentList): Get the dimension names for an ExperimentList using CharacterList

colnames(ExperimentList): Get the column names for an ExperimentList as a CharacterList slightly more efficiently

rownames(ExperimentList): Get the row names for an ExperimentList as a CharacterList slightly more efficiently

mergeReplicates(ExperimentList): Apply the mergeReplicates method on the ExperimentList elements

assay(x = ANY, i = missing): Obtain the specified assay with a numeric or character reference

assays(ExperimentList): Get the assay data from each element in the ExperimentList

coercion

Convert a list or S4 List to an ExperimentList using the as() function.

In the following example, x is either a list or List:

    as(x, "ExperimentList")

Examples

ExperimentList()
#> ExperimentList class object of length 0:
#>