MatchedAssayExperiment - A matched-samples MultiAssayExperiment class

This class supports the use of matched samples where an equal number of observations per biological unit are present in all assays.

Usage

MatchedAssayExperiment(...)

Arguments

...

Either a single MultiAssayExperiment or the components to create a valid MultiAssayExperiment

Value

A MatchedAssayExperiment object

Functions

MatchedAssayExperiment(): Construct a MatchedAssayExperiment class from MultiAssayExperiment

See also

Examples

data("miniACC")
acc <- as(miniACC, "MatchedAssayExperiment")
acc
#> A MatchedAssayExperiment object of 5 listed
#>  experiments with user-defined names and respective classes.
#>  Containing an ExperimentList class object of length 5:
#>  [1] RNASeq2GeneNorm: SummarizedExperiment with 198 rows and 43 columns
#>  [2] gistict: SummarizedExperiment with 198 rows and 43 columns
#>  [3] RPPAArray: SummarizedExperiment with 33 rows and 43 columns
#>  [4] Mutations: matrix with 97 rows and 43 columns
#>  [5] miRNASeqGene: SummarizedExperiment with 471 rows and 43 columns
#> Functionality:
#>  experiments() - obtain the ExperimentList instance
#>  colData() - the primary/phenotype DataFrame
#>  sampleMap() - the sample coordination DataFrame
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment
#>  *Format() - convert into a long or wide DataFrame
#>  assays() - convert ExperimentList to a SimpleList of matrices
#>  exportClass() - save data to flat files