MultiAssayExperiment: Build an integrative multi-assay container

MultiAssayExperiment allows the manipulation of related multiassay datasets with partially overlapping samples, associated metadata at the level of an entire study, and at the level of the "biological unit". The biological unit may be a patient, plant, yeast strain, etc.

Details

The package hierarchy of information:

study

experiments

samples

See also

Author

Maintainer: Marcel Ramos marcel.ramos@roswellpark.org (ORCID)

Authors:

Martin Morgan [contributor]

Vincent J Carey [contributor]

Levi Waldron lwaldron.research@gmail.com

Other contributors:

Lori Shepherd [contributor]

Hervé Pagès [contributor]

MultiAssay SIG biocmultiassay@googlegroups.com [contributor]