Checking assay method for any class

The hasAssay function is intended for developers who would like to include new classes into a MultiAssayExperiment instance. It checks the methods tables of the assay function for the specified class of the argument.

hasAssay(object)

Arguments

object

A MultiAssayExperiment or named list object instance

Value

A logical value indicating method availability

Examples

lst <- structure(list(), .Names=character())
hasAssay(lst)
#> [1] TRUE