Adrenocortical Carcinoma (ACC) MultiAssayExperiment

A MultiAssayExperiment object providing a reduced version of the TCGA ACC dataset for all 92 patients. RNA-seq, copy number, and somatic mutations are included only for genes whose proteins are included in the reverse-phase protein array. The MicroRNA-seq dataset is also included, with infrequently expressed microRNA removed. Clinical, pathological, and subtype information are provided by colData(miniACC), and some additional details are provided by metadata(miniACC).

Usage

data("miniACC")

Format

A MultiAssayExperiment with 5 experiments, providing:

RNASeq2GeneNorm

RNA-seq count data: an ExpressionSet with 198 rows and 79 columns

gistict

Reccurent copy number lesions identified by GISTIC2: a SummarizedExperiment with 198 rows and 90 columns

RPPAArray

Reverse Phase Protein Array: an ExpressionSet with 33 rows and 46 columns. Rows are indexed by genes, but protein annotations are available from featureData(miniACC[["RPPAArray"]]). The source of these annotations is noted in abstract(miniACC[["RPPAArray"]])

Mutations

Somatic mutations: a matrix with 223 rows and 90 columns. 1 for any kind of non-silent mutation, zero for silent (synonymous) or no mutation.

miRNASeqGene

microRNA sequencing: an ExpressionSet with 471 rows and 80 columns. Rows not having at least 5 counts in at least 5 samples were removed.

Source

https://github.com/waldronlab/multiassayexperiment-tcga

References

Zheng S et al.: Comprehensive Pan-Genomic Characterization of Adrenocortical Carcinoma. Cancer Cell 2016, 29:723-736.

Author

Levi Waldron lwaldron.research@gmail.com

Examples

data("miniACC")
metadata(miniACC)
#> $title
#> [1] "Comprehensive Pan-Genomic Characterization of Adrenocortical Carcinoma"
#> 
#> $PMID
#> [1] "27165744"
#> 
#> $sourceURL
#> [1] "http://s3.amazonaws.com/multiassayexperiments/accMAEO.rds"
#> 
#> $RPPAfeatureDataURL
#> [1] "http://genomeportal.stanford.edu/pan-tcga/show_target_selection_file?filename=Allprotein.txt"
#> 
#> $colDataExtrasURL
#> [1] "http://www.cell.com/cms/attachment/2062093088/2063584534/mmc3.xlsx"
#> 
colnames(colData(miniACC))
#>  [1] "patientID"                           
#>  [2] "years_to_birth"                      
#>  [3] "vital_status"                        
#>  [4] "days_to_death"                       
#>  [5] "days_to_last_followup"               
#>  [6] "tumor_tissue_site"                   
#>  [7] "pathologic_stage"                    
#>  [8] "pathology_T_stage"                   
#>  [9] "pathology_N_stage"                   
#> [10] "gender"                              
#> [11] "date_of_initial_pathologic_diagnosis"
#> [12] "radiation_therapy"                   
#> [13] "histological_type"                   
#> [14] "residual_tumor"                      
#> [15] "number_of_lymph_nodes"               
#> [16] "race"                                
#> [17] "ethnicity"                           
#> [18] "Histology"                           
#> [19] "C1A.C1B"                             
#> [20] "mRNA_K4"                             
#> [21] "MethyLevel"                          
#> [22] "miRNA.cluster"                       
#> [23] "SCNA.cluster"                        
#> [24] "protein.cluster"                     
#> [25] "COC"                                 
#> [26] "OncoSign"                            
#> [27] "purity"                              
#> [28] "ploidy"                              
#> [29] "genome_doublings"                    
#> [30] "ADS"                                 
table(miniACC$vital_status)
#> 
#>  0  1 
#> 58 34 
longFormat(
    miniACC["MAPK3", , ],
    colDataCols = c("vital_status", "days_to_death")
)
#> harmonizing input:
#>   removing 216 sampleMap rows not in names(experiments)
#> DataFrame with 169 rows and 7 columns
#>               assay      primary     rowname                      colname
#>         <character>  <character> <character>                     <factor>
#> 1   RNASeq2GeneNorm TCGA-OR-A5J1       MAPK3 TCGA-OR-A5J1-01A-11R-A29S-07
#> 2   RNASeq2GeneNorm TCGA-OR-A5J2       MAPK3 TCGA-OR-A5J2-01A-11R-A29S-07
#> 3   RNASeq2GeneNorm TCGA-OR-A5J3       MAPK3 TCGA-OR-A5J3-01A-11R-A29S-07
#> 4   RNASeq2GeneNorm TCGA-OR-A5J5       MAPK3 TCGA-OR-A5J5-01A-11R-A29S-07
#> 5   RNASeq2GeneNorm TCGA-OR-A5J6       MAPK3 TCGA-OR-A5J6-01A-31R-A29S-07
#> ...             ...          ...         ...                          ...
#> 165         gistict TCGA-PA-A5YG       MAPK3 TCGA-PA-A5YG-01A-11D-A29H-01
#> 166         gistict TCGA-PK-A5H9       MAPK3 TCGA-PK-A5H9-01A-11D-A29H-01
#> 167         gistict TCGA-PK-A5HA       MAPK3 TCGA-PK-A5HA-01A-11D-A29H-01
#> 168         gistict TCGA-PK-A5HB       MAPK3 TCGA-PK-A5HB-01A-11D-A29H-01
#> 169         gistict TCGA-PK-A5HC       MAPK3 TCGA-PK-A5HC-01A-11D-A309-01
#>         value vital_status days_to_death
#>     <numeric>    <integer>     <integer>
#> 1     946.681            1          1355
#> 2    1699.382            1          1677
#> 3    2347.243            0            NA
#> 4    1299.156            1           365
#> 5    2543.424            0            NA
#> ...       ...          ...           ...
#> 165         1            0            NA
#> 166         0            0            NA
#> 167         0            0            NA
#> 168         0            0            NA
#> 169         1            0            NA

wideFormat(
    miniACC["MAPK3", , ],
    colDataCols = c("vital_status", "days_to_death")
)
#> harmonizing input:
#>   removing 216 sampleMap rows not in names(experiments)
#> DataFrame with 92 rows and 5 columns
#>          primary vital_status days_to_death RNASeq2GeneNorm_MAPK3 gistict_MAPK3
#>      <character>    <integer>     <integer>             <numeric>     <numeric>
#> 1   TCGA-OR-A5J1            1          1355               946.681             0
#> 2   TCGA-OR-A5J2            1          1677              1699.382             0
#> 3   TCGA-OR-A5J3            0            NA              2347.243             1
#> 4   TCGA-OR-A5J4            1           423                    NA             0
#> 5   TCGA-OR-A5J5            1           365              1299.156             1
#> ...          ...          ...           ...                   ...           ...
#> 88  TCGA-PK-A5H9            0            NA               1013.47             0
#> 89  TCGA-PK-A5HA            0            NA               2446.12             0
#> 90  TCGA-PK-A5HB            0            NA               1463.31             0
#> 91  TCGA-PK-A5HC            0            NA                    NA             1
#> 92  TCGA-P6-A5OG            1           383               1310.88            -1