A container for multi-experiment data

The ExperimentList class is a container that builds on the SimpleList with additional checks for consistency in experiment names and length. It contains a SimpleList of experiments with sample identifiers. One element present per experiment performed.

# S4 method for ExperimentList
show(object)

# S4 method for ExperimentList
isEmpty(x)

# S4 method for ExperimentList
dimnames(x)

# S4 method for ExperimentList
mergeReplicates(x, replicates = list(),
  simplify = BiocGenerics::mean, ...)

# S4 method for ANY,missing
assay(x, i, ...)

# S4 method for ExperimentList
assays(x, ..., withDimnames = TRUE)

# S4 method for ExperimentList,missing
assay(x, i, ...)

# S4 method for ExperimentList,numeric
assay(x, i, ...)

# S4 method for ExperimentList,character
assay(x, i, ...)

Arguments

object, x
An ExperimentList object

replicates
mergeReplicates: A list or LogicalList where each element represents a sample and a vector of repeated measurements for the sample

simplify
A function for merging columns where duplicates are indicated by replicates

...
Additional arguments. See details for more information.

i
A scalar character or integer index

withDimnames
logical (default TRUE) whether to return dimension names

Details

Convert from SimpleList or list to the multi-experiment data container. When using the mergeReplicates method, additional arguments are passed to the given simplify function argument (e.g., na.rm = TRUE)

Methods (by generic)

show: Show method for ExperimentList class

isEmpty: check for zero length across all experiments

dimnames: Get the dimension names for an ExperimentList using CharacterList

mergeReplicates: Apply the mergeReplicates method on the ExperimentList elements

assay: Obtain the specified assay with a numberic or character reference

assays: Get the assay data from each element in the ExperimentList

coercion

Convert a list or S4 List to an ExperimentList using the as() function.

In the following example, x is either a list or List:

    \code{as(x, "ExperimentList")}

Examples

ExperimentList()

#> ExperimentList class object of length 0: