Accessing/modifying slot information

A set of accessor and setter generic functions to extract either the sampleMap, the ExperimentList, colData, or metadata slots of a MultiAssayExperiment object

# S4 method for MultiAssayExperiment
sampleMap(x)

# S4 method for MultiAssayExperiment
experiments(x)

# S4 method for MultiAssayExperiment
colData(x, ...)

# S4 method for MultiAssayExperiment
metadata(x)

# S4 method for MultiAssayExperiment,DataFrame
sampleMap(object) <- value

  # S4 method for MultiAssayExperiment,ExperimentList
experiments(object) <- value

# S4 method for MultiAssayExperiment,DataFrame
colData(x) <- value

# S4 method for MultiAssayExperiment
metadata(x, ...) <- value

# S4 method for MultiAssayExperiment
$(x, name) <- value

# S4 method for MultiAssayExperiment
names(x) <- value

# S4 method for MultiAssayExperiment
$(x, name)

Arguments

x
A MultiAssayExperiment object

...
Argument not in use

object
A MultiAssayExperiment object

value
See details.

name
A column in colData

Value

Accessors: Either a sampleMap, ExperimentList, or DataFrame object

Setters: A MultiAssayExperiment object

Accessors

Eponymous names for accessing MultiAssayExperiment slots with the exception of the ExperimentList accessor named experiments.

colData: Access the colData slot

sampleMap: Access the sampleMap slot

experiments: Access the ExperimentList slot

`[[`: Access the ExperimentList slot

`$`: Access a column in colData

Setters

Setter method values (i.e., 'function(x) <- value'):

experiments<-: An ExperimentList object containing experiment data of supported classes

sampleMap<-: A DataFrame object relating samples to biological units and assays

colData<-: A DataFrame object describing the biological units

metadata<-: A list object of metadata

`[[<-`: Equivalent to the experiments<- setter method for convenience

`$<-`: A vector to replace the indicated column in colData

Examples

## Load example MultiAssayExperiment
example(MultiAssayExperiment)

#> Warning: no help found for ‘MultiAssayExperiment’

## Access the sampleMap
sampleMap(mae)

#> Error in sampleMap(mae): object 'mae' not found

## Replacement method for a MultiAssayExperiment sampleMap
sampleMap(mae) <- S4Vectors::DataFrame()

#> Error in sampleMap(mae) <- S4Vectors::DataFrame(): object 'mae' not found

## Access the ExperimentList
experiments(mae)

#> Error in experiments(mae): object 'mae' not found

## Replace with an empty ExperimentList
experiments(mae) <- ExperimentList()

#> Error in experiments(mae) <- ExperimentList(): object 'mae' not found

## Access the metadata
metadata(mae)

#> Error in metadata(mae): object 'mae' not found

## Replace metadata with a list
metadata(mae) <- list(runDate =
    format(Sys.time(), "%B %d, %Y"))

#> Error in metadata(mae) <- list(runDate = format(Sys.time(), "%B %d, %Y")): object 'mae' not found

## Access the colData
colData(mae)

#> Error in colData(mae): object 'mae' not found

## Access a column in colData
mae$age

#> Error in eval(expr, envir, enclos): object 'mae' not found

## Replace a column in colData
mae$age <- mae$age + 1

#> Error in eval(expr, envir, enclos): object 'mae' not found