Adrenocortical Carcinoma (ACC) MultiAssayExperiment

A MultiAssayExperiment object providing a reduced version of the TCGA ACC dataset for all 92 patients. RNA-seq, copy number, and somatic mutations are included only for genes whose proteins are included in the reverse-phase protein array. The MicroRNA-seq dataset is also included, with infrequently expressed microRNA removed. Clinical, pathological, and subtype information are provided by colData(miniACC), and some additional details are provided by metadata(miniACC).

miniACC

Format

A MultiAssayExperiment with 5 experiments, providing:

RNASeq2GeneNorm

RNA-seq count data: an ExpressionSet with 198 rows and 79 columns

gistict

Reccurent copy number lesions identified by GISTIC2: a SummarizedExperiment with 198 rows and 90 columns

RPPAArray

Reverse Phase Protein Array: an ExpressionSet with 33 rows and 46 columns. Rows are indexed by genes, but protein annotations are available from featureData(miniACC[["RPPAArray"]]). The source of these annotations is noted in abstract(miniACC[["RPPAArray"]])

Mutations

Somatic mutations: a matrix with 223 rows and 90 columns. 1 for any kind of non-silent mutation, zero for silent (synonymous) or no mutation.

miRNASeqGene

microRNA sequencing: an ExpressionSet with 471 rows and 80 columns. Rows not having at least 5 counts in at least 5 samples were removed.

Source

https://github.com/waldronlab/multiassayexperiment-tcga

References

Zheng S *et al.*: Comprehensive Pan-Genomic Characterization of Adrenocortical Carcinoma. Cancer Cell 2016, 29:723-736.

Examples

miniACC

#> A MultiAssayExperiment object of 5 listed
#>  experiments with user-defined names and respective classes. 
#>  Containing an ExperimentList class object of length 5: 
#>  [1] RNASeq2GeneNorm: SummarizedExperiment with 198 rows and 79 columns 
#>  [2] gistict: SummarizedExperiment with 198 rows and 90 columns 
#>  [3] RPPAArray: SummarizedExperiment with 33 rows and 46 columns 
#>  [4] Mutations: matrix with 97 rows and 90 columns 
#>  [5] miRNASeqGene: SummarizedExperiment with 471 rows and 80 columns 
#> Features: 
#>  experiments() - obtain the ExperimentList instance 
#>  colData() - the primary/phenotype DataFrame 
#>  sampleMap() - the sample availability DataFrame 
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment 
#>  *Format() - convert into a long or wide DataFrame 
#>  assays() - convert ExperimentList to a SimpleList of matrices

metadata(miniACC)

#> $title
#> [1] "Comprehensive Pan-Genomic Characterization of Adrenocortical Carcinoma"
#> 
#> $PMID
#> [1] "27165744"
#> 
#> $sourceURL
#> [1] "http://s3.amazonaws.com/multiassayexperiments/accMAEO.rds"
#> 
#> $RPPAfeatureDataURL
#> [1] "http://genomeportal.stanford.edu/pan-tcga/show_target_selection_file?filename=Allprotein.txt"
#> 
#> $colDataExtrasURL
#> [1] "http://www.cell.com/cms/attachment/2062093088/2063584534/mmc3.xlsx"
#> 

colnames(colData(miniACC))

#>  [1] "patientID"                           
#>  [2] "years_to_birth"                      
#>  [3] "vital_status"                        
#>  [4] "days_to_death"                       
#>  [5] "days_to_last_followup"               
#>  [6] "tumor_tissue_site"                   
#>  [7] "pathologic_stage"                    
#>  [8] "pathology_T_stage"                   
#>  [9] "pathology_N_stage"                   
#> [10] "gender"                              
#> [11] "date_of_initial_pathologic_diagnosis"
#> [12] "radiation_therapy"                   
#> [13] "histological_type"                   
#> [14] "residual_tumor"                      
#> [15] "number_of_lymph_nodes"               
#> [16] "race"                                
#> [17] "ethnicity"                           
#> [18] "Histology"                           
#> [19] "C1A.C1B"                             
#> [20] "mRNA_K4"                             
#> [21] "MethyLevel"                          
#> [22] "miRNA.cluster"                       
#> [23] "SCNA.cluster"                        
#> [24] "protein.cluster"                     
#> [25] "COC"                                 
#> [26] "OncoSign"                            
#> [27] "purity"                              
#> [28] "ploidy"                              
#> [29] "genome_doublings"                    
#> [30] "ADS"                                 

table(miniACC$vital_status)

#> 
#>  0  1 
#> 58 34 

longFormat(miniACC["MAPK3", , ], colDataCols = c("vital_status", "days_to_death"))

#> harmonizing input:
#>   removing 216 sampleMap rows not in names(experiments)

#> DataFrame with 169 rows and 7 columns
#>               assay      primary     rowname                      colname
#>         <character>  <character> <character>                  <character>
#> 1   RNASeq2GeneNorm TCGA-OR-A5J1       MAPK3 TCGA-OR-A5J1-01A-11R-A29S-07
#> 2   RNASeq2GeneNorm TCGA-OR-A5J2       MAPK3 TCGA-OR-A5J2-01A-11R-A29S-07
#> 3   RNASeq2GeneNorm TCGA-OR-A5J3       MAPK3 TCGA-OR-A5J3-01A-11R-A29S-07
#> 4   RNASeq2GeneNorm TCGA-OR-A5J5       MAPK3 TCGA-OR-A5J5-01A-11R-A29S-07
#> 5   RNASeq2GeneNorm TCGA-OR-A5J6       MAPK3 TCGA-OR-A5J6-01A-31R-A29S-07
#> ...             ...          ...         ...                          ...
#> 165         gistict TCGA-PA-A5YG       MAPK3 TCGA-PA-A5YG-01A-11D-A29H-01
#> 166         gistict TCGA-PK-A5H9       MAPK3 TCGA-PK-A5H9-01A-11D-A29H-01
#> 167         gistict TCGA-PK-A5HA       MAPK3 TCGA-PK-A5HA-01A-11D-A29H-01
#> 168         gistict TCGA-PK-A5HB       MAPK3 TCGA-PK-A5HB-01A-11D-A29H-01
#> 169         gistict TCGA-PK-A5HC       MAPK3 TCGA-PK-A5HC-01A-11D-A309-01
#>         value vital_status days_to_death
#>     <numeric>    <integer>     <integer>
#> 1    946.6811            1          1355
#> 2   1699.3821            1          1677
#> 3   2347.2434            0            NA
#> 4   1299.1557            1           365
#> 5   2543.4239            0            NA
#> ...       ...          ...           ...
#> 165         1            0            NA
#> 166         0            0            NA
#> 167         0            0            NA
#> 168         0            0            NA
#> 169         1            0            NA

wideFormat(miniACC["MAPK3", , ], colDataCols = c("vital_status", "days_to_death"))

#> harmonizing input:
#>   removing 216 sampleMap rows not in names(experiments)

#> DataFrame with 92 rows and 5 columns
#>          primary vital_status days_to_death RNASeq2GeneNorm_MAPK3 gistict_MAPK3
#>      <character>    <integer>     <integer>             <numeric>     <numeric>
#> 1   TCGA-OR-A5J1            1          1355              946.6811             0
#> 2   TCGA-OR-A5J2            1          1677             1699.3821             0
#> 3   TCGA-OR-A5J3            0            NA             2347.2434             1
#> 4   TCGA-OR-A5J4            1           423                    NA             0
#> 5   TCGA-OR-A5J5            1           365             1299.1557             1
#> ...          ...          ...           ...                   ...           ...
#> 88  TCGA-PK-A5H9            0            NA             1013.4694             0
#> 89  TCGA-PK-A5HA            0            NA             2446.1246             0
#> 90  TCGA-PK-A5HB            0            NA             1463.3079             0
#> 91  TCGA-PK-A5HC            0            NA                    NA             1
#> 92  TCGA-P6-A5OG            1           383             1310.8799            -1

##
## The following is the code used to create this mini dataset from the full ACC dataset.
## The full ACC MultiAssayExperiment was created by the pipeline at
## https://github.com/waldronlab/multiassayexperiment-tcga.

# NOT RUN {
    ## See www.tinyurl.com/MAEOurls for more pre-built TCGA MultiAssayExperiment objects
    download.file("http://s3.amazonaws.com/multiassayexperiments/accMAEO.rds",
                  destfile = "accMAEO.rds")
    library(MultiAssayExperiment)
    library(RaggedExperiment) #needed for RaggedExperiment objects by updateObject()
    library(Biobase)

    acc <- readRDS("accMAEO.rds")
    acc <- updateObject(acc)
    protmap <- read.csv(paste0("http://genomeportal.stanford.edu/",
                        "pan-tcga/target_selection_send_data",
                        "?filename=Allprotein.txt"), as.is = TRUE
        )

    RPPAgenes <- Filter(function(x) x != "", protmap$Genes)
    RPPAgenes <- unlist(strsplit(RPPAgenes, ","))
    RPPAgenes <- unique(RPPAgenes)

    miniACC <-
        acc[RPPAgenes, , c("RNASeq2GeneNorm", "gistict", "RPPAArray", "Mutations")]
    mut <- assay(miniACC[["Mutations"]], i = "Variant_Classification")
    mut <- ifelse(is.na(mut) | mut == "Silent", 0, 1)

    miniACC[["Mutations"]] <- mut
    colData(miniACC) <- colData(miniACC)[, c(1:17, 810:822)]

    rpparowData <-
        protmap[match(rownames(miniACC[["RPPAArray"]]), protmap$Genes),]
    rpparowData <- AnnotatedDataFrame(rpparowData)
    featureData(miniACC[["RPPAArray"]]) <- rpparowData

    md <-
        list(
            title = "Comprehensive Pan-Genomic Characterization of Adrenocortical Carcinoma",
            PMID = "27165744",
            sourceURL = "http://s3.amazonaws.com/multiassayexperiments/accMAEO.rds",
            RPPAfeatureDataURL = paste0("http://genomeportal.stanford.edu/",
                                        "pan-tcga/show_target_selection_file",
                                        "?filename=Allprotein.txt"),
            colDataExtrasURL = "http://www.cell.com/cms/attachment/2062093088/2063584534/mmc3.xlsx"
        )
    metadata(miniACC) <- md

    mirna <- acc[["miRNASeqGene"]]
    mirna <- mirna[rowSums(assay(mirna) >= 5) >= 5, ]
    experimentData(mirna)@abstract <-
        "Note: Rows not having at least 5 counts in at least 5 samples were removed."
    miniACC <- c(miniACC,
                 list(miRNASeqGene = mirna),
                 sampleMap = sampleMap(acc)[sampleMap(acc)$assay == "miRNASeqGene",])

     miniACC[["RNASeq2GeneNorm"]] <-
         as(miniACC[["RNASeq2GeneNorm"]], "SummarizedExperiment")
     miniACC[["RPPAArray"]] <-
         as(miniACC[["RPPAArray"]], "SummarizedExperiment")
     miniACC[["miRNASeqGene"]] <-
         as(miniACC[["miRNASeqGene"]], "SummarizedExperiment")

    save(miniACC, file = "data/miniACC.RData", compress = "bzip2")
# }