Subsetting a MultiAssayExperiment object

A set of functions for extracting and dividing a MultiAssayExperiment

subsetByRow(x, y, ...)

subsetByColData(x, y)

subsetByColumn(x, y)

subsetByAssay(x, y)

# S4 method for ExperimentList,ANY
subsetByRow(x, y, ...)

# S4 method for ExperimentList,list
subsetByRow(x, y)

# S4 method for ExperimentList,List
subsetByRow(x, y)

# S4 method for ExperimentList,logical
subsetByRow(x, y)

# S4 method for ExperimentList,list
subsetByColumn(x, y)

# S4 method for ExperimentList,List
subsetByColumn(x, y)

# S4 method for ExperimentList,logical
subsetByColumn(x, y)

# S4 method for ExperimentList
subsetByAssay(x, y)

# S4 method for MultiAssayExperiment,ANY
subsetByColData(x, y)

# S4 method for MultiAssayExperiment,character
subsetByColData(x, y)

# S4 method for MultiAssayExperiment,ANY
subsetByRow(x, y, ...)

# S4 method for MultiAssayExperiment,ANY
subsetByColumn(x, y)

# S4 method for MultiAssayExperiment
subsetByAssay(x, y)

# S4 method for MultiAssayExperiment,ANY,ANY,ANY
[(x, i, j, k, ...,
  drop = TRUE)

# S4 method for MultiAssayExperiment,ANY,ANY
[[(x, i, j, ...)

# S4 method for MultiAssayExperiment,ANY,ANY
[[(x, i, j, ...) <- value

Arguments

x
A MultiAssayExperiment or ExperimentList

y
Any argument used for subsetting, can be a character, logical, integer, list or List vector

...
Additional arguments passed on to lower level functions.

i
Either a character, integer, logical or GRanges object for subsetting by rows

j
Either a character, logical, or numeric vector for subsetting by colData rows. See details for more information.

k
Either a character, logical, or numeric vector for subsetting by assays

drop
logical (default TRUE) whether to drop empty assay elements in the ExperimentList

value
An assay compatible with the MultiAssayExperiment API

Details

Subsetting a MultiAssayExperiment by the j index can yield a call to either subsetByColData or subsetByColumn. For vector inputs, the subset will be applied to the colData rows. For List-type inputs, the List will be applied to each of the elements in the ExperimentList. The order of the subsetting elements in the List must match that of the ExperimentList in the MultiAssayExperiment.

subsetBycolData: Select biological units by vector input types

subsetByColumn: Select observations by assay or for each assay

subsetByRow: Select rows by assay or for each assay

subsetByAssay: Select experiments

Examples

## Load the example MultiAssayExperiment
example("MultiAssayExperiment")

#> Warning: no help found for ‘MultiAssayExperiment’

## Using experiment names
subsetByAssay(mae, "Affy")

#> Error in subsetByAssay(mae, "Affy"): object 'mae' not found

## Using numeric indices
subsetByAssay(mae, 1:2)

#> Error in subsetByAssay(mae, 1:2): object 'mae' not found

## Using a logical vector
subsetByAssay(mae, c(TRUE, FALSE, TRUE))

#> Error in subsetByAssay(mae, c(TRUE, FALSE, TRUE)): object 'mae' not found

## Subset by character vector (Jack)
subsetByColData(mae, "Jack")

#> Error in subsetByColData(mae, "Jack"): object 'mae' not found

## Subset by numeric index of colData rows (Jack and Bob)
subsetByColData(mae, c(1, 3))

#> Error in subsetByColData(mae, c(1, 3)): object 'mae' not found

## Subset by logical indicator of colData rows (Jack and Jill)
subsetByColData(mae, c(TRUE, TRUE, FALSE, FALSE))

#> Error in subsetByColData(mae, c(TRUE, TRUE, FALSE, FALSE)): object 'mae' not found

subsetByColumn(mae, list(Affy = 1:2,
    Methyl450k = c(3,5,2), RNASeqGene = 2:4, GISTIC = 1))

#> Error in subsetByColumn(mae, list(Affy = 1:2, Methyl450k = c(3, 5, 2),     RNASeqGene = 2:4, GISTIC = 1)): object 'mae' not found

subsetWith <- S4Vectors::mendoapply(`[`, colnames(mae),
    MoreArgs = list(1:2))

#> Error in colnames(mae): object 'mae' not found

subsetByColumn(mae, subsetWith)

#> Error in subsetByColumn(mae, subsetWith): object 'mae' not found

## Use a GRanges object to subset rows where ranged data present
egr <- GenomicRanges::GRanges(seqnames = "chr2",
    IRanges::IRanges(start = 11, end = 13), strand = "-")
subsetByRow(mae, egr)

#> Error in subsetByRow(mae, egr): object 'mae' not found

## Use a logical vector (recycling used)
subsetByRow(mae, c(TRUE, FALSE))

#> Error in subsetByRow(mae, c(TRUE, FALSE)): object 'mae' not found

## Use a character vector
subsetByRow(mae, "ENST00000355076")

#> Error in subsetByRow(mae, "ENST00000355076"): object 'mae' not found